Fixing merge conflicts

+++ six/modules/c++/six.sicd/include/six/sicd/SlantPlanePixelTransformer.h
@@ -41,7 +41,7 @@ namespace sicd
  *  \class SlantPlanePixelTransformer
  *  \brief Projects a slant plane pixel into various ground plane coordinates
  *
- *  This class should be used to project coordinates between slant and 
+ *  This class should be used to project coordinates between slant and
  *  ground planes.
  */
 class SlantPlanePixelTransformer
@@ -65,21 +65,21 @@ public:
      *  \param pixel - Slant Plane pixel with (row,col) index
      *  \return      - Returns the ground plane location in ECEF
      */
-    scene::Vector3 toECEF(const types::RowCol<size_t>& pixel) const;
+    scene::Vector3 toECEF(const types::RowCol<double>& pixel) const;
 
     /*!
      *  \fn toLLA
      *  \param pixel - Slant Plane pixel with (row,col) index
      *  \return      - Returns the ground plane location in LLA
      */
-    scene::LatLonAlt toLLA(const types::RowCol<size_t>& pixel) const;
+    scene::LatLonAlt toLLA(const types::RowCol<double>& pixel) const;
 
     /*!
      *  \fn toLatLon
      *  \param pixel - Slant Plane pixel with (row,col) index
      *  \return      - Returns the ground plane location in LatLon
      */
-    scene::LatLon toLatLon(const types::RowCol<size_t>& pixel) const;
+    scene::LatLon toLatLon(const types::RowCol<double>& pixel) const;
 
 private:
     const scene::SceneGeometry mGeom;
